Moved nnotations to more global namespace
